-_- zzzZZZZZZzZZZZZzZZZZ
git-svn-id: https://www.etud.insa-toulouse.fr/svn/projet_tut_OS/trunk/tacos@366 ab7442c9-e926-469a-b309-6eee04cae01a

+++ include/sem.h
@@ -5,9 +5,9 @@
 
 int semget(uint8_t key);
 int semcreate(uint8_t key);
-int semdel(uint8_t semid);
-int semP(uint8_t semid);
-int semV(uint8_t semid);
+int semdel(uint32_t semid);
+int semP(uint32_t semid);
+int semV(uint32_t semid);
 
 #endif //SEM_H
